requirements.txt
asgiref==3.7.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==4.2
django-admin-persian-fonts==0.2
django-admin-volt==1.0.10
django-ckeditor==6.7.1
django-jazzmin==2.6.1
django-js-asset==2.2.0
django-mysql==4.12.0
django-parler==2.3
django-rosetta==0.10.0
django-taggit==5.0.1
django-tinymce==4.0.0
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
parler==1.0.1
pillow==10.2.0
polib==1.2.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
scipy==1.13.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
